chore(wire): register G1.B Tauri bridges + fix rust 1.95 clippy regressions
Wires the 3 new crates' Tauri commands into the desktop shell:
  apohara-verification: quality_gates_evaluate
  apohara-safety:       safety_{check_permission, analyze_bash_compound, match_pattern}
  apohara-spec:         spec_{load_plan, get_plan_status}

PlanStatusCache registered as managed state (Arc) so spec_get_plan_status
can borrow it via tauri::State without reparsing every call.

Side fix: rust 1.95 introduces 4 new clippy lints (manual_contains,
manual_div_ceil, manual_is_multiple_of, unnecessary_cast) that fire on
pre-existing code in apohara-client/ssh-server/remote-worker. Inline
1-line fixes keep cargo clippy --workspace clean for the cierre gate.

+++ crates/apohara-client/src/connect.rs
@@ -45,7 +45,7 @@ impl BackoffPolicy {
         let offset = clock.jitter_ms(jitter_window);
         // Apply jitter symmetrically: capped ± up-to jitter_window.
         let final_ms = if offset.is_negative() {
-            capped.saturating_sub(offset.unsigned_abs() as u64)
+            capped.saturating_sub(offset.unsigned_abs())
         } else {
             capped.saturating_add(offset as u64)
         };
